Fix Fraction(-2147483648.0) for 32-bit wide long

+++ tools/source/generic/fract.cxx
@@ -136,7 +136,7 @@ Fraction::Fraction( double dVal )
         return;
     }
 
-    while ( std::abs( (long)dVal ) < nMAX && nDen < nMAX )
+    while ( std::abs( dVal ) < nMAX && nDen < nMAX )
     {
         dVal *= 10;
         nDen *= 10;
